move map_ids to part_tables
